Reorganize project files into proper packages
Currently, the project files were all located under src/main/java,
which makes it hard to navigate code and less modular.

command package for all command-related classes.
task package for task-related classes.
exception package for custom exceptions.
ui package for UI-related classes.

Add checkstyle configuration.
