+++ src/main/java/exception/InvalidArgumentException.java
@@ -1,3 +1,4 @@
+package exception;
 public class InvalidArgumentException extends SylvieException {
     public InvalidArgumentException(String message) {
         super(message);
