+++ src/main/java/exception/SylvieException.java
@@ -1,3 +1,4 @@
+package exception;
 public class SylvieException extends Exception {
     public SylvieException(String message) {
         super(message);
