+++ src/main/java/task/Task.java
@@ -1,3 +1,5 @@
+package task;
+
 public class Task {
     private final String description;
     private boolean isDone;
